feat: #918 moving towards automatic refresh
