+++ src/a22_plan_viewer/test/test_world_flask_app.py
@@ -10,15 +10,15 @@ def test_planviewer_IsCreated():
     response = client.get("/")  # or some valid route you define
 
     # WHEN / THEN
-    # If '/' route not defined, skip or test /process with dummy data
+    # If '/' route not defined, skip or test /refresh_tree with dummy data
     assert response.status_code in (200, 404)
 
 
-def test_process_endpoint():
+def test_refresh_tree_endpoint():
     # ESTABLISH
     client = planviewer.test_client()
     response = client.post(
-        "/process", json={"_mode": "default", "data": {"foo": "bar"}}
+        "/refresh_tree", json={"_mode": "default", "data": {"foo": "bar"}}
     )
     assert response.status_code == 200
